backend/requirements/common.txt
numpy
Pillow
asgiref
django
djangorestframework
libsass 
django-compressor 
django-sass-processor
django-debug-toolbar
pytz
mypy
django-stub

pymongo==3.12.1
django-seed==0.3.1
gunicorn==20.0.4
python-dotenv==0.13.0
sqlparse==0.3.1
mysqlclient
argon2-cffi